Renamed files from Arcball to Orbig
I think it fits much better.
